% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{get_bathy}
\alias{get_bathy}
\title{Retrieve bathymetric data}
\usage{
get_bathy(utm30n_extent, extent_mult, bathy_levels)
}
\arguments{
\item{utm30n_extent}{An \code{extent} object in UTM 30N coordinates (e.g. the
extent of the Northern Gannet tracks)}

\item{extent_mult}{A multiplier applied to the extent (e.g. 5)}

\item{bathy_levels}{A vector of bathymetric contour levels (e.g. \code{c(-50, -150, -250)})}
}
\value{
An \code{sf} object of the bathymetric contours
}
\description{
Retrieve bathymetric data
}
