% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{get_naturalearth}
\alias{get_naturalearth}
\title{Retrieve Natural Earth data}
\usage{
get_naturalearth(ne_type)
}
\arguments{
\item{ne_type}{Type of Natural Earth data e.g. "land" or "minor islands"}
}
\value{
An \code{sf} object
}
\description{
Retrieve Natural Earth data
}
