% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dist.R
\name{home_range}
\alias{home_range}
\title{Extract home ranges at 25\%, 50\%, 75\%, and 95\% utilization distribution levels}
\usage{
home_range(uds)
}
\arguments{
\item{uds}{Adult and immature utilization distributions}
}
\value{
An sf data.frame with columns: age ("adult" or "immature"), area,
ud_level (a factor: 25\%, 50\%, 75\%, and 95\%), and geometry,
}
\description{
Extract home ranges at 25\%, 50\%, 75\%, and 95\% utilization distribution levels
}
