% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fig1.R
\name{make_fig1}
\alias{make_fig1}
\title{Make figure 1}
\usage{
make_fig1(home_ranges, land, islands, bathy)
}
\arguments{
\item{home_ranges}{Northern Gannet home ranges (see \code{home_range})}

\item{land}{\code{sf} object with land geometries (see \code{get_naturalearth})}

\item{islands}{\code{sf} object with minor island geometries (see
\code{get_naturalearth})}

\item{bathy}{\code{sf} object with bathymetric contours (see \code{get_bathy})}
}
\value{
\code{ggplot} object of figure 1
}
\description{
Figure 1 is a map comparing the home ranges of adult and immature Northern
Gannets.
}
