% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fig2.R
\name{make_fig2}
\alias{make_fig2}
\title{Make figure 2}
\usage{
make_fig2(obs_ba, null_ba)
}
\arguments{
\item{obs_ba}{Observed Bhattacharyya's affinity of adult and immature home
ranges (see \code{observed_ba})}

\item{null_ba}{Null distribution of Bhattacharyya's affinity of adult and
immature home ranges (see \code{sim_null_ba})}
}
\value{
File path to figure 2
}
\description{
Figure 2 shows the observed home range overlap of adult and immature Northern
Gannets compared the null distribution
}
