% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table1.R
\name{make_table1}
\alias{make_table1}
\title{Make table 1}
\usage{
make_table1(trips)
}
\arguments{
\item{trips}{see \code{trip_summaries}}
}
\value{
a data frame with four rows (trip metrics: duration, length, range,
and area) and 9 columns (metric name; median, min, and max of metric by age
category; chisquared statistic and p-value for likelihood ratio test).
}
\description{
Table 1 is a summary of foraging trip metrics
}
