% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya.R
\name{observed_ba}
\alias{observed_ba}
\title{Observed home range overlap between adult and immature Northern Gannets}
\usage{
observed_ba(uds)
}
\arguments{
\item{uds}{Northern Gannet utilization distribution (see \code{util_dist})}
}
\value{
A data frame with variables: percent (50\%, 95\%) and ba
(Bhattacharyya's affinity for 50\% and 95\% UD home ranges)
}
\description{
Observed home range overlap between adult and immature Northern Gannets
}
