% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya.R
\name{sim_null_ba}
\alias{sim_null_ba}
\title{Null distribution of home range overlap between adult and immature Northern
Gannets}
\usage{
sim_null_ba(n_cores = 1)
}
\arguments{
\item{n_cores}{Number of cores}
}
\value{
A data frame with variables: percent (50\%, 95\%), ba (Bhattacharyya's
affinity for 50\% and 95\% UD home ranges), and replicate (1:1000)
}
\description{
Calculated by shuffling adult and immature age assignments.
}
