% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_table1.R
\name{trip_summaries}
\alias{trip_summaries}
\title{Summarize trips}
\usage{
trip_summaries()
}
\value{
A data with 247 rows and 8 variables: age category, individual id,
trip id, trip duration (hours), trip departure angle (radians), trip length
(km), trip range (km), and trip area (km^2)
}
\description{
Summarize trips
}
