% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util_dist.R
\name{util_dist}
\alias{util_dist}
\title{Utilization distribution of tracks}
\usage{
util_dist(tracks)
}
\arguments{
\item{tracks}{Use \code{all_tracks_df} for the actual data, or shuffle ages to generate null Bhattacharyya's affinity distributions}
}
\value{
A list of two \code{estUD} objects: adult and immature
}
\description{
Utilization distribution of tracks
}
